backend/document-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-multipart==0.0.18
httpx==0.27.2
